Update Modal Storybook examples to use doc site example
